move JsonApiIdentifiable interface to wrapper package
