--- models/src/main/java/org/jvalue/ods/api/jsonapi/JsonApiIdentifiableWithMetaData.java
@@ -1,5 +1,0 @@
-package org.jvalue.ods.api.jsonapi;
-
-public interface JsonApiIdentifiableWithMetaData extends JsonApiIdentifiable {
-	MetaData getMetaData();
-}
