--- models/src/main/java/org/jvalue/ods/api/jsonapi/MetaData.java
@@ -1,4 +1,0 @@
-package org.jvalue.ods.api.jsonapi;
-
-public interface MetaData {
-}
