+++ server/src/main/java/org/jvalue/ods/rest/v2/jsonapi/document/JsonApiResource.java
@@ -3,7 +3,7 @@ package org.jvalue.ods.rest.v2.jsonapi.document;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.jvalue.ods.api.jsonapi.JsonApiIdentifiable;
+import org.jvalue.ods.rest.v2.jsonapi.wrapper.JsonApiIdentifiable;
 import org.jvalue.ods.utils.JsonUtils;
 
 import java.net.URI;
