+++ server/src/main/java/org/jvalue/ods/utils/JsonUtils.java
@@ -3,7 +3,7 @@ package org.jvalue.ods.utils;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import org.jvalue.ods.api.jsonapi.JsonApiIdentifiable;
+import org.jvalue.ods.rest.v2.jsonapi.wrapper.JsonApiIdentifiable;
 
 import java.util.LinkedList;
 import java.util.List;
