+++ server/src/test/java/org/jvalue/ods/rest/v2/jsonapi/response/TestEntityProvider.java
@@ -1,7 +1,7 @@
 package org.jvalue.ods.rest.v2.jsonapi.response;
 
-import org.jvalue.ods.api.jsonapi.JsonApiIdentifiable;
 import org.jvalue.ods.rest.v2.jsonapi.document.JsonApiResourceIdentifier;
+import org.jvalue.ods.rest.v2.jsonapi.wrapper.JsonApiIdentifiable;
 
 import java.util.HashMap;
 import java.util.Map;
