+++ server/src/test/java/org/jvalue/ods/utils/JsonUtilsTest.java
@@ -3,7 +3,7 @@ package org.jvalue.ods.utils;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.junit.Assert;
 import org.junit.Test;
-import org.jvalue.ods.api.jsonapi.JsonApiIdentifiable;
+import org.jvalue.ods.rest.v2.jsonapi.wrapper.JsonApiIdentifiable;
 import org.jvalue.ods.rest.v2.jsonapi.document.JsonApiResourceIdentifierTest;
 
 public class JsonUtilsTest {
